* fixed madvise/fadvise detection * fixed many compiler warnings -- Peter Breitenlohner peb mppmu.mpg.de
git-svn-id: svn://svn.oetiker.ch/rrdtool/trunk/program@1269 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_cgi.c
@@ -1430,7 +1430,7 @@ s_var   **rrdcgiReadVariables(
 
             /* try to find out if there's already such a variable */
             for (k = 0; k < i && (strncmp(result[k]->name, cp, esp - cp)
-                                  || !(strlen(result[k]->name) == esp - cp));
+                                  || !(strlen(result[k]->name) == (size_t)(esp - cp)));
                  k++);
 
             if (k == i) {   /* No such variable yet */
